Lots of work
Implement win state + rematch
Add OverlayGUI for full-screen GUI elements
Add DespawnAll to ObjectPool for post-game cleanup
Add GameOver methods to classes that need to do work post-game

Disable all unnecessary shadow casting
Move to shader-based terrain texturing
Add TerrainTextureOffset class for randomizing texture offset
Add TerrainTypeTexture for matching worley noise distance metrics to
terrain textures

+++ Assets/Scripts/AvatarGUI.cs
@@ -14,8 +14,8 @@ public class AvatarGUI : MonoBehaviour
 	int activeDroneIndex = 0;
 	List<Vector2> droneIconPositions;
 	List<Vector2> droneIconTargets;
-	int health = 0;
-	int healthTarget;
+	float health = 0;
+	float healthTarget;
 	float dash = 0;
 	float dashTarget;
 	float prevRespawn = 0f;
@@ -103,8 +103,8 @@ public class AvatarGUI : MonoBehaviour
 				}
 
 				// Health
-				int healthTarget = target.parent.GetComponent<Avatar>().Health;
-				health = (int)Mathf.Lerp( health, healthTarget, BarLerpFactor * Time.deltaTime );
+				float healthTarget = target.parent.GetComponent<Avatar>().Health;
+				health = Mathf.Lerp( health, healthTarget, BarLerpFactor * Time.deltaTime );
 				GUIContent healthText = new GUIContent( "Health" );
 				GUIContent healthNumber = new GUIContent( health.ToString() );
